Introduce sequential transaction runner
For emulator contention resolution
Update deps
