+++ package.json
@@ -29,13 +29,14 @@
   "dependencies": {
     "@google-cloud/firestore": "^4.8.1",
     "@types/pino": "^6.3.5",
-    "fs-extra": "^9.0.1",
+    "fs-extra": "^9.1.0",
     "lodash": "^4.17.20",
-    "strapi-utils": "^3.4.3"
+    "p-queue": "^6.6.2",
+    "strapi-utils": "^3.4.4"
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.6",
-    "@types/lodash": "^4.14.167",
+    "@types/lodash": "^4.14.168",
     "@types/node": "^12.19.13",
     "rimraf": "^3.0.2",
     "typescript": "^4.1.3"
