+++ test/package.json
@@ -6,21 +6,21 @@
     "coverage-combine": "istanbul-combine -d ../coverage/combined -p detail -r json ../coverage/*/coverage.json"
   },
   "dependencies": {
-    "firebase-tools": "^9.2.1",
+    "firebase-tools": "^9.2.2",
     "jest": "^26.6.3",
     "lodash": "^4.17.20",
     "patch-package": "^6.2.2",
     "qs": "^6.9.6",
     "remap-istanbul": "^0.13.0",
     "request-promise-native": "^1.0.9",
-    "strapi": "^3.4.3",
-    "strapi-admin": "^3.4.3",
+    "strapi": "^3.4.4",
+    "strapi-admin": "^3.4.4",
     "strapi-connector-firestore": "file:..",
-    "strapi-plugin-content-manager": "^3.4.3",
-    "strapi-plugin-content-type-builder": "^3.4.3",
-    "strapi-plugin-upload": "^3.4.3",
-    "strapi-plugin-users-permissions": "^3.4.3",
-    "strapi-utils": "^3.4.3",
+    "strapi-plugin-content-manager": "^3.4.4",
+    "strapi-plugin-content-type-builder": "^3.4.4",
+    "strapi-plugin-upload": "^3.4.4",
+    "strapi-plugin-users-permissions": "^3.4.4",
+    "strapi-utils": "^3.4.4",
     "supertest": "^6.1.1",
     "wait-on": "^5.2.1"
   },
@@ -28,7 +28,7 @@
     "@actions/core": "^1.2.6",
     "@actions/github": "^4.0.0",
     "@actions/io": "^1.0.2",
-    "fs-extra": "^9.0.1",
+    "fs-extra": "^9.1.0",
     "glob": "^7.1.6",
     "glob-promise": "^3.4.0",
     "istanbul-combine": "^0.3.0"
